+++ tests/conftest.py
@@ -16,6 +16,7 @@ from typer.testing import CliRunner as TyperCliRunner
 
 import rich_click.rich_click as rc
 from rich_click._compat_click import CLICK_IS_BEFORE_VERSION_82
+from rich_click._compat_typer import TYPER_IS_BEFORE_VERSION_026
 from rich_click.rich_command import RichCommand
 from rich_click.rich_context import RichContext
 
@@ -119,7 +120,7 @@ def cli_runner() -> CliRunner:
 
 @pytest.fixture
 def typer_cli_runner() -> TyperCliRunner:
-    if CLICK_IS_BEFORE_VERSION_82:
+    if CLICK_IS_BEFORE_VERSION_82 and TYPER_IS_BEFORE_VERSION_026:
         return TyperCliRunner(mix_stderr=False)  # type: ignore[call-arg,unused-ignore]
     else:
         return TyperCliRunner()