+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Version 1.9.8 (2026-05-28)
 
-- Fix regression in patching with Click 8.4.0.
+- Fixed regression in patching with Click 8.4.0.
+- Fixed regression in patching with Typer 0.26.0.
 
 ## Version 1.9.7 (2026-01-30)